feat: create fromJson toJson  ProjectRequestModel
