+++ front/lib/app/domain/presentation/project/screen/project_update.dart
@@ -1,11 +1,11 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:front/app/domain/presentation/project/component/project_manager_selector.dart';
 import 'package:front/app/domain/presentation/project/component/project_from.dart';
+import 'package:front/app/domain/presentation/project/component/project_manager_selector.dart';
 import 'package:front/app/domain/presentation/project/viewmodel/project.dart';
 import 'package:front/core/const/enum.dart';
-import 'package:front/core/project/data/models/project_create.dart';
+import 'package:front/core/project/data/models/project_request.dart';
 import 'package:front/core/project/domain/entities/project.dart';
 import 'package:front/core/user/models/user.dart';
 import 'package:front/shared/atom/bottom_navigation_bar.dart';
