+++ front/lib/core/project/data/repositories/project_repository_impl.dart
@@ -3,8 +3,7 @@ import 'dart:io';
 import 'package:dartz/dartz.dart';
 import 'package:front/core/project/data/data_sources/remote_data_source.dart';
 import 'package:front/core/project/data/data_sources/temp_data_source.dart';
-import 'package:front/core/project/data/models/project.dart';
-import 'package:front/core/project/data/models/project_create.dart';
+import 'package:front/core/project/data/models/project_request.dart';
 import 'package:front/core/project/domain/entities/project.dart';
 import 'package:front/core/project/domain/repositories/project_repository.dart';
 import 'package:front/core/utils/exception.dart';
