+++ front/lib/core/project/domain/usecases/create_project_usecase.dart
@@ -1,5 +1,5 @@
 import 'package:dartz/dartz.dart';
-import 'package:front/core/project/data/models/project_create.dart';
+import 'package:front/core/project/data/models/project_request.dart';
 import 'package:front/core/project/domain/entities/project.dart';
 import 'package:front/core/project/domain/repositories/project_repository.dart';
 import 'package:front/core/utils/failure.dart';
